+++ front/src/widgets/Progressbar/index.ts
@@ -1,0 +1,1 @@
+export { Progressbar } from "./ui/Progressbar/Progressbar";
