+++ front/src/widgets/Progressbar/model/type/progressbar.ts
@@ -1,0 +1,9 @@
+export const ProgressSteps: { id: string, title: string }[] = [
+  { 'id': '1', 'title': 'Новая' },
+  { 'id': '2', 'title': 'Обрабатывается' },
+  { 'id': '3', 'title': 'На согласовании у заказчика' },
+  { 'id': '4', 'title': 'Стоимость работ' },
+  { 'id': '5', 'title': 'В работе' },
+  { 'id': '6', 'title': 'На согласовании у заказчика' },
+  { 'id': '7', 'title': 'Сделано' },
+];
