+++ front/src/widgets/Progressbar/ui/Progressbar/Progressbar.module.scss
