+++ front/src/widgets/Progressbar/ui/Progressstep/Progressstep.module.scss
